#24 rename: ChartsPanel pages 폴더에서 components 폴더로 이동
